requirements.txt
# Web Framework
Flask==2.0.2
flask-cors==6.0.1
Werkzeug==2.0.2
Jinja2==3.0.1
flask-socketio==5.1.1
python-socketio==5.4.0
python-engineio==4.3.1
eventlet==0.33.0
dnspython==2.2.1
simple-websocket==1.1.0

# Database
mysql-connector-python==8.0.32
python-dotenv==0.19.0
SQLAlchemy==1.4.41

# Data Processing and Analysis
pandas==1.5.3
pandas_ta
numpy==1.23.5
yfinance==0.2.18

# Technical Analysis (Enhanced Trading)
ta-lib==0.4.28
scipy==1.10.1
scikit-learn==1.3.0
matplotlib==3.7.1
seaborn==0.12.2

# API and HTTP
requests==2.31.0
httpx==0.24.1
alpha_vantage==2.3.1

# Angel One SmartAPI (Official Requirements)
smartapi-python==1.4.8
pyotp==2.8.0
logzero==1.7.0
websocket-client==1.8.0
pycryptodome
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==43.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isodate==0.6.1
pycparser==2.21
pyparsing==3.1.0
rdflib==6.3.2
rdflib-jsonld==0.6.2
simplejson==3.19.1
six==1.16.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
urllib3==2.0.3
zope.interface==6.0

# PDF Generation
reportlab==3.6.12

# Date and Time
python-dateutil==2.8.2
pytz==2023.3

# Security
cryptography==43.0.1
PyJWT==2.10.1

# Utilities
tqdm==4.64.1
setuptools==65.5.1
schedule==1.2.2

# Testing
pytest==7.2.0
pytest-cov==4.0.0

# Development Tools
black==22.12.0
flake8==6.0.0